refactor: separate styled files + grey color not configured
