+++ src/components/organisms/PageHeader/PageHeader.tsx
@@ -29,7 +29,8 @@ import DiscordLogo from '@assets/DiscordLogo.svg';
 import MonokleKubeshopLogo from '@assets/MonokleKubeshopLogo.svg';
 
 import ClusterSelection from './ClusterSelection';
-import * as S from './Styled';
+
+import * as S from './styled';
 
 const ExitButton = (props: {onClick: () => void}) => {
   const {onClick} = props;
