+++ src/components/organisms/PageHeader/ProjectSelection.tsx
@@ -27,7 +27,7 @@ import FileExplorer from '@components/atoms/FileExplorer';
 
 import {useFileExplorer} from '@hooks/useFileExplorer';
 
-import * as S from './Styled';
+import * as S from './ProjectSelection.styled';
 
 const ProjectSelection = () => {
   const dispatch = useAppDispatch();
@@ -111,22 +111,23 @@ const ProjectSelection = () => {
               />
             </Tooltip>
             <Tooltip mouseEnterDelay={TOOLTIP_DELAY} title={NewEmptyProjectTooltip} placement="bottomRight">
-              <S.ProjectFolderAddOutlined onClick={() => handleCreateProject(false)} />
+              <S.FolderAddOutlined onClick={() => handleCreateProject(false)} />
             </Tooltip>
             <Tooltip mouseEnterDelay={TOOLTIP_DELAY} title={NewProjectFromTemplateTooltip} placement="bottomRight">
-              <S.ProjectFormatPainterOutlined onClick={() => handleCreateProject(true)} />
+              <S.FormatPainterOutlined onClick={() => handleCreateProject(true)} />
             </Tooltip>
           </S.ProjectsMenuActionsContainer>
         </S.ProjectsMenuContainer>
-        <S.ProjectsTable
+
+        <S.Table
           size="small"
           showSorterTooltip={false}
           dataSource={searchText ? filteredProjects : projects}
           pagination={false}
           scroll={{y: 300}}
           rowKey="rootFolder"
-          onRow={(project: Project) => ({
-            onClick: () => handleProjectChange(project),
+          onRow={project => ({
+            onClick: () => handleProjectChange(project as Project),
           })}
         >
           <Column
@@ -202,28 +203,28 @@ const ProjectSelection = () => {
               </S.ProjectTableActions>
             )}
           />
-        </S.ProjectsTable>
+        </S.Table>
       </S.ProjectMenu>
     );
   };
 
   return (
     <S.ProjectsDropdown
-      isClusterSelectorVisible={isClusterSelectorVisible}
+      $isClusterSelectorVisible={isClusterSelectorVisible}
+      arrow
       overlay={projectMenu}
       placement="bottomCenter"
-      arrow
       trigger={['click']}
-      onVisibleChange={setIsDropdownMenuVisible}
       visible={isDropdownMenuVisible}
+      onVisibleChange={setIsDropdownMenuVisible}
     >
-      <Tooltip mouseEnterDelay={TOOLTIP_DELAY} title={ProjectManagementTooltip} placement="bottomRight">
-        <S.ProjectButton>
+      <Tooltip mouseEnterDelay={TOOLTIP_DELAY} placement="bottomRight" title={ProjectManagementTooltip}>
+        <S.Button>
           <S.FolderOpenOutlined />
           <S.ProjectName>{activeProject?.name}</S.ProjectName>
           <S.DownOutlined />
           <FileExplorer {...fileExplorerProps} />
-        </S.ProjectButton>
+        </S.Button>
       </Tooltip>
     </S.ProjectsDropdown>
   );
